test(node:worker_threads): address CodeRabbit review feedback - Move the undici regression test to test/regression/issue/29423.test.ts   (per repo convention for issue-linked tests) and drop the stale TODO. - In markAsUncloneable.test.ts, build the child worker's file:// URL with   pathToFileURL(join(dir, "worker.mjs")).href in the parent test scope and   inject it via JSON.stringify, so paths with spaces or a Windows drive   letter work correctly and are properly percent-encoded. - Add `if (exitCode !== 0) expect(stderr).toBe("")` before the final   exitCode assertion in the child-worker test so a non-zero exit surfaces   the full stderr diff in CI logs. - Bump the 29423 regression test's per-test timeout to 120s via the 3rd   argument to test(...), so a cold-cache `bun install undici@^8.1.0` has   room to either succeed or hit its existing "install failed → skip"   branch instead of dying at the default 5s wall. Verified locally with `bun bd test`: - markAsUncloneable.test.ts: 25/25 pass (incl. child-worker at 2210ms) - test/regression/issue/29423.test.ts: 1/1 pass (5921ms, real npm install) No implementation or behavior changes.
